Update "replace existing" logic, use config directly
The old logic was "If file exists AND option is unchecked, skip replacing. Otherwise replace." Now it's "If file doesn't exist OR option is checked, replace. Otherwise skip." Not sure why the old one wouldn't work in a specific case, but this way seems extra safe.
Also using the settings object now for checking these conditions rather than whether the checkboxes are checked or not.
Bumped version to 1.2.1.